Move python test maps from 'good_maps' directory
If the CirclesDatasource class was not available from the python_plugin_tests module then the 'load_good_maps' test
would fail. The class would not be available if, for example, the python plugin had not been built or if shapely was not
available on the test system. (Pull request #1414 removes the dependency on Shapely.)

The error reporting should be tidied up for this case but for the moment, move the Python plugin's test maps into their
own directory since they're not guaranteed to be 'good maps' in all cases at the moment.

Hopefully addresses issue #1407