Added visual test for multi-line label alignment.
Calculation of label alignment in multi-line labels seems to be
off. This test reproduces it by laying out a series of multi-line
labels along an S-shaped curve. In all these cases, the vertical
bar character should, to within a few pixels, line up. The
orientation at sharper corners is still a little off, but likely
due to discontinuities.