Wrapped GDAL errors in Mapnik handler.
This means it's easy to silence all or some of the reporting from
GDAL. Without this, all messages were handled by the default
handler, which printed them to `stdout`. That's probably the
behaviour that most people want, except it isn't wanted during
testing when we're testing for expected failures - then it's just
garbage interfering with reading the test output.