Add test for XML parser whitespace trimming behaviour.
This tests the at the XML parser trims whitespace from XML text
nodes. This was already the behaviour of the libxml2 parser, but
not the rapidxml parser.

Note that this makes #2878 pass for the rapidxml parser as well
as the libxml2 parser. It seems that Travis uses the rapidxml
parser only.