Don't trim whitespace in the XML loaders.
The libxml2 and ptree (a.k.a rapidxml) XML loaders were
inconsistent in their handling of whitespace:

* Rapidxml trimmed some, but not did not process CDATA child
  nodes.
* Libxml2 trimmed all, due to manual application of the trim
  function.

As discussed in https://github.com/mapnik/mapnik/pull/2878#issuecomment-108728845
it would be preferable not to trim. This causes test failures due
to unexpected whitespace, but these are confined to the text node
formatting and CSV plugin's inline functionality.

If a manual call to trim is made for CSV inline handling and
all-whitespace child nodes are dropped in the text node formatting,
then all the tests pass for me. Note that we still don't have full
test coverage, so it's possible that other behaviour may have
changed.