Fix failing tests on Mac OS X.
The issue here was that the relative offset overload of
`basic_istream::seekg` cannot, according to the C++11 standard,
move away from EOF. This was fixed for the absolute offset
overload in the C++11 standard, and implemented in libc++ and
libstdc++. However, libstdc++ also implements it for the relative
overload. This will be fixed in a future version of the standard by
http://www.open-std.org/jtc1/sc22/wg21/docs/lwg-active.html#2244

In the CSV input plugin, however, we only move away from EOF when
rewinding to the very beginning of the file, so using the absolute
offset overload instead fixes the problem.