Revert "benchmark - slightly better stderr"
This partially reverts commit 5ab6db26079c5b51bc920392eb3a75b4242a2347.

The change of argument --threads N doesn't just change output, it
changes the mode of operation.
With --threads 0, the benchmark runs wholly in the main thread.
With --threads 1, it starts 1 worker thread, which does the work, and
the main thread measures how long it takes to complete.